Feat: refactor logic
Minorly refactors logic that could be safely simplified to do the
same behavior as the provided logic.

Adds a docstring in 'joust.py' as a bonus to replace a comment but
mainly helps me understand at a glance what the function does
and ways it may be improved at a later date.

Premature optimization may be the death of innovation but I like
to make incremental improvements when I'm bored while I think of
the next features.

Author:    Elliot Jones <64710849+ellietjones@users.noreply.github.com>